Refatora testes de Cadastro e Login para o component de LoginFormComponent.
