+++ src/pages/cadastro/Cadastro.test.js
@@ -13,16 +13,9 @@ describe('Testes de Cadastro', () => {
     });
 
     it('Deve renderizar componente corretamente', () => {
-      const usuarioInput = screen.getByRole('textbox', { name: 'Usuário' });
-      const senhaInput = screen.getByText('Senha');
-      const confirmarSenhaInput = screen.getByText('Confirmar senha');
-      const botaoCadastrar = screen.getByRole('button', { name: 'Cadastrar' });
+      const tituloCadastro = screen.getByText('Cadastro');
 
-      expect(usuarioInput).toBeInTheDocument();
-      expect(senhaInput).toBeInTheDocument();
-      expect(confirmarSenhaInput).toBeInTheDocument();
-
-      expect(botaoCadastrar).toBeInTheDocument();
+      expect(tituloCadastro).toBeInTheDocument();
     });
 
     it('Deve exibir link para a página de Login', () => {
@@ -31,80 +24,6 @@ describe('Testes de Cadastro', () => {
       });
       expect(linkLogin).toBeInTheDocument();
       expect(linkLogin).toHaveAttribute('href', '/login');
-    });
-  });
-
-  describe('Testes de habilitar/desabilitar botão de Cadastro', () => {
-    beforeEach(() => {
-      render(<Cadastro />);
-    });
-
-    describe('Deve desabilitar botão de Cadastrar quando', () => {
-      it('Usuário for inválido', () => {
-        const usuario = screen.getByTestId('usuario');
-        const senha = screen.getByTestId('senha');
-
-        fireEvent.change(usuario, {
-          target: { value: 'ale' },
-        });
-
-        fireEvent.change(senha, { target: { value: 'thefamemonster' } });
-        expect(
-          screen.getByRole('button', { name: 'Cadastrar' })
-        ).toBeDisabled();
-      });
-
-      it('Senha for inválida', () => {
-        const usuario = screen.getByTestId('usuario');
-        const senha = screen.getByTestId('senha');
-
-        fireEvent.change(usuario, {
-          target: { value: 'alejandro' },
-        });
-
-        fireEvent.change(senha, { target: { value: 'the' } });
-        expect(
-          screen.getByRole('button', { name: 'Cadastrar' })
-        ).toBeDisabled();
-      });
-
-      it('Senhas forem diferentes', () => {
-        const usuario = screen.getByTestId('usuario');
-        const senha = screen.getByTestId('senha');
-        const confirmarSenha = screen.getByTestId('confirmarSenha');
-
-        fireEvent.change(usuario, {
-          target: { value: 'alejandro' },
-        });
-
-        fireEvent.change(senha, { target: { value: 'thefamemonster' } });
-        fireEvent.change(confirmarSenha, { target: { value: 'thefame' } });
-
-        expect(
-          screen.getByRole('button', { name: 'Cadastrar' })
-        ).toBeDisabled();
-      });
-    });
-
-    describe('Deve habilitar o botão de Cadastrar quando', () => {
-      it('Os campos estiverem preenchidos corretamente ', () => {
-        const usuario = screen.getByTestId('usuario');
-        const senha = screen.getByTestId('senha');
-        const confirmarSenha = screen.getByTestId('confirmarSenha');
-
-        fireEvent.change(usuario, {
-          target: { value: 'alejandro' },
-        });
-
-        fireEvent.change(senha, { target: { value: 'thefamemonster' } });
-        fireEvent.change(confirmarSenha, {
-          target: { value: 'thefamemonster' },
-        });
-
-        expect(
-          screen.getByRole('button', { name: 'Cadastrar' })
-        ).not.toBeDisabled();
-      });
     });
   });
